/base/branches/1.2.2 /include/util/buffers/CycleBuffers.h修正编译警告
git-svn-id: http://svn01/repositories/r01/base/branches/1.2.2@5079 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ buffers/CycleBuffers.h
@@ -748,7 +748,7 @@ namespace util
 
             void check()
             {
-                assert(write_.position() - read_.position() <= capacity_);
+                assert((size_t)(write_.position() - read_.position()) <= capacity_);
                 if (read_.begin_remainder == write_.begin_remainder) {
                     assert(this->egptr() == this->pbase() || this->epptr() == this->eback());
                 }
